Removing unnecessary casting
- Also changing scope of getString method in AldeJsonObjectWrapper.
- Correcting a few comments as well.

+++ src/main/java/eu/tango/self/adaptation/manager/model/ApplicationDeployment.java
@@ -55,7 +55,7 @@ public class ApplicationDeployment extends AldeJsonObjectWrapper {
      */
     public int getExecutableId() {
         if (json.has("executable_id")) {
-            return (int) json.getInt("executable_id");
+            return json.getInt("executable_id");
         }
         //the default assumption is zero.
         return 0;        
@@ -67,7 +67,7 @@ public class ApplicationDeployment extends AldeJsonObjectWrapper {
      */
     public int getTestbedId() {
         if (json.has("testbed_id")) {
-            return (int) json.getInt("testbed_id");
+            return json.getInt("testbed_id");
         }
         //the default assumption is zero.
         return 0;
